Add comparison script
For comparing Python 2/3 CSV diffs. Only change should be precision.

We assume that the only the first value in the CSV row changes and that
the rest of the line remains unchanged. This seemed reasonable based on
scanning the diff by eye (and this script shows that this is the case).

This script only works with Python 3.5+ due to using math.isclose().